+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Fixed
+- Support data source files that in are not in the root folder
+
 ## [2.0.2] - 2021-08-31
 
 ### Fixed